rw/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
googletrans==4.0.0rc1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.4.2
msgpack==1.0.8
multidict==6.0.5
nltk==3.8.1
openai==0.28.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pymongo==4.8.0
pyparsing==3.1.2
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
